+++ src/lib/diffHandler.js
@@ -32,7 +32,7 @@ class DiffHandler {
         this.namelessClasses = namelessClasses.slice();
 
         // Although we may be the source of truth for 'hostname', we do not want
-        // to diff it beccause hostname is set in 2 different places. Better
+        // to diff it because hostname is set in 2 different places. Better
         // to let f5-cloud-libs handle checking it.
         for (let i = 0; i < this.classesOfTruth.length; i += 1) {
             if (this.classesOfTruth[i] === 'hostname') {